ADD: Separate folder for net things
