+++ kernel/net/net.c
@@ -1,0 +1,9 @@
+#include "kernel/net/arp.h"
+#include "kernel/net/ip.h"
+
+void
+net_init()
+{
+    arp_init();
+    ip_init();
+}
